Make the `FTDataBuffer` a lock-free data structure
It's now a single-producer, single-consumer, lock-free data structure.
This is possible due to python's GIL that assures that only one single
thread writes at a specific time. We only needed to take care of data
consistency.
It's also a good base if we need to switch to using processes instead of
threads.